Flatten slash defer into interaction-errors; export withChatLock for tests
Drop the one-call-site slash-ack module and the withChatLockForTests pass-throughs.
