+++ apps/platform/telegram/src/chat-lock.test.ts
@@ -2,7 +2,7 @@ import { afterEach, describe, expect, test } from "bun:test";
 import {
   resetChatLocksForTests,
   seedChatLockForTests,
-  withChatLockForTests,
+  withChatLock,
 } from "./chat-handler";
 
 afterEach(() => {
@@ -24,7 +24,7 @@ describe("withChatLock rejection safety", () => {
       seedChatLockForTests("chat:reject", stale);
 
       let ran = false;
-      await withChatLockForTests("chat:reject", async () => {
+      await withChatLock("chat:reject", async () => {
         ran = true;
       });
       await Bun.sleep(20);
